KYC/requirement.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python>=4.8.0
easyocr>=1.7.1